test: extract overlay animation styles tests into separate file
Separate animation behavior tests from animation CSS assertions, as
done in the notification package. Also fix a comment that referenced
the removed `animated-styles.js`.
